Major update of SQL Server labs

+++ SQLServerLab/create_VMs.ps1
@@ -12,23 +12,29 @@ foreach ($computerName in $vmConfig.Keys) {
     Write-PSFMessage -Level Host -Message "Creating virtual maschine $computerName"
     New-MyAzureLabVM -ComputerName $computerName -SourceImage $vmConfig.$computerName.SourceImage -VMSize $vmConfig.$computerName.VMSize -Credential $initCredential -TrustedLaunch -EnableException
 }
-# The following also sets $statusConfig.Uri and $domainConfig.DCIPAddress
-Write-PSFMessage -Level Host -Message "Creating virtual maschine STATUS"
-New-MyAzureLabStatusVM -EnableException
+if ($Env:MyStatusURL) {
+    $statusConfig.Uri = $Env:MyStatusURL
+    $domainConfig.DCIPAddress = (Get-AzNetworkInterface -ResourceGroupName $resourceGroupName -Name "DC_Interface").IpConfigurations[0].PrivateIpAddress
+} else {
+    # The following also sets $statusConfig.Uri and $domainConfig.DCIPAddress
+    Write-PSFMessage -Level Host -Message "Creating virtual maschine STATUS"
+    New-MyAzureLabStatusVM -EnableException
+}
 
 
 ##########
 Write-PSFMessage -Level Host -Message 'Part 2: Setting up the active directory domain'
 ##########
 
-# Renaming the virtual maschines
+# Setting up CredSSP and WinRM
 # Installing software
 # Setting up PowerShell
 # Installing PowerShell modules
 # Setting up domain
+# Creating AD users on domain controller
 # Setting up file server on domain controller
 
-$partStartedAt = [datetime]::Now
+$partStartedAt = ([datetime]::Now).ToUniversalTime()
 foreach ($computerName in $vmConfig.Keys) {
     Write-PSFMessage -Level Host -Message "Configuring virtual maschine $computerName"
     Invoke-MyAzureLabDeployment -ComputerName $computerName -Credential $initCredential -Path $vmConfig.$computerName.Script_A -Config $vmConfig.$computerName -EnableException
@@ -42,11 +48,9 @@ Wait-MyAzureLabDeploymentCompletion -OnlyStatusAfter $partStartedAt -EnableExcep
 Write-PSFMessage -Level Host -Message 'Part 3: Setting up SQL Server resources'
 ##########
 
-# Creating AD users
 # Filling file server with sql server sources
-# Setting up CredSSP
 
-$partStartedAt = [datetime]::Now
+$partStartedAt = ([datetime]::Now).ToUniversalTime()
 foreach ($computerName in $vmConfig.Keys) {
     if ($vmConfig.$computerName.Script_B) {
         Write-PSFMessage -Level Host -Message "Configuring virtual maschine $computerName"
@@ -62,7 +66,7 @@ Wait-MyAzureLabDeploymentCompletion -OnlyStatusAfter $partStartedAt -EnableExcep
 Write-PSFMessage -Level Host -Message 'Part 4: Setting up SQL Server instances'
 ##########
 
-$partStartedAt = [datetime]::Now
+$partStartedAt = ([datetime]::Now).ToUniversalTime()
 foreach ($computerName in $vmConfig.Keys) {
     if ($vmConfig.$computerName.Script_C) {
         Write-PSFMessage -Level Host -Message "Configuring virtual maschine $computerName"
@@ -79,8 +83,10 @@ Write-PSFMessage -Level Host -Message "Waiting 2 minutes"
 Start-Sleep -Seconds 120
 Wait-MyAzureLabDeploymentCompletion -OnlyStatusAfter $partStartedAt -EnableException
 
-Write-PSFMessage -Level Host -Message "Removing virtual maschine STATUS"
-Remove-MyAzureLabVM -ComputerName STATUS -EnableException
+if (-not $Env:MyStatusURL) {
+    Write-PSFMessage -Level Host -Message "Removing virtual maschine STATUS"
+    Remove-MyAzureLabVM -ComputerName STATUS -EnableException
+}
 
 $deploymentDuration = [datetime]::Now - $deploymentStart
 Write-PSFMessage -Level Host -Message "Finished deployment after $([int]$deploymentDuration.TotalMinutes) minutes"
